Removed unused imports

+++ app/api/seed/route.ts
@@ -1,13 +1,12 @@
 import {IDS_TRACON_FACILITIES} from "@/facility/facilities";
-import {TraconConfig} from "@/types";
 import prisma from "@/lib/db";
 
 export const dynamic = "force-dynamic";
 export async function GET() {
 
-    // if ((await prisma.tracon.findMany()).length > 0) {
-    //     return Response.json("Database has already been seeded.  If you have changed the configuration, make sure all the previous data is deleted.");
-    // }
+    if ((await prisma.tracon.findMany()).length > 0) {
+        return Response.json("Database has already been seeded.  If you have changed the configuration, make sure all the previous data is deleted.");
+    }
 
     await prisma.tracon.deleteMany();
     for (const tracon of IDS_TRACON_FACILITIES) {
